Implementation of decay through a resistor
