+++ SundialsWrapper.cpp
@@ -1,40 +1,5 @@
 
-#include <arkode/arkode_arkstep.h>     /* access to ARKode func., consts. */
-#include <nvector/nvector_serial.h>    /* access to serial N_Vector       */
-#include <sunmatrix/sunmatrix_dense.h> /* access to dense SUNMatrix       */
-#include <sunlinsol/sunlinsol_dense.h> /* access to dense SUNLinearSolver */
-#include <sundials/sundials_types.h>   /* defs. of realtype, sunindextype */
-
-// Number of equations/unknowns in system
-#define N_DIMENSIONS 2 
-
-#define ION_TEMP_IDX 0
-#define ELEC_TEMP_IDX 1
-#define DENSITY_IDX 2
-#define MOMENTUM_IDX 2 
-
-#define MACH_IDX 1
-
-#define ION_TEMPERATURE( u ) NV_Ith_S( u, ION_TEMP_IDX )
-#define ELECTRON_TEMPERATURE( u ) NV_Ith_S( u, ELEC_TEMP_IDX )
-#define DENSITY( u ) NV_Ith_S( u, DENSITY_IDX )
-#define VOLTAGE( u ) NV_Ith_S( u, MOMENTUM_IDX )
-
-#define MACH_NUMBER( u ) NV_Ith_S( u, MACH_IDX )
-
-#define ION_HEAT_BALANCE( F ) NV_Ith_S( F, ION_TEMP_IDX )
-#define ELECTRON_HEAT_BALANCE( F ) NV_Ith_S( F, ELEC_TEMP_IDX )
-#define PARTICLE_BALANCE( F ) NV_Ith_S( F, DENSITY_IDX )
-#define MOMENTUM_BALANCE( F ) NV_Ith_S( F, MOMENTUM_IDX )
-
-#define IRK_SCHEME static_cast<ARKODE_DIRKTableID>( ARKSTEP_DEFAULT_DIRK_4 )
-#define ARKSTEP_NULL_STEPPER  static_cast<ARKODE_ERKTableID>( -1 )
-
-#include "Config.hpp"
-#include "MirrorPlasma.hpp"
-#include <exception>
-#include <iostream>
-#include <iomanip>
+#include "SundialsWrapper.hpp"
 
 void ArkodeErrorWrapper( int errorFlag, std::string&& fName )
 {
@@ -116,88 +81,6 @@ int ARKStep_TemperatureSolve( realtype t, N_Vector u, N_Vector uDot, void* voidP
 	return ARK_SUCCESS;
 }
 
-int ARKStep_FreeWheel( realtype t, N_Vector u, N_Vector uDot, void* voidPlasma )
-{
-	MirrorPlasma* plasmaPtr = reinterpret_cast<MirrorPlasma*>( voidPlasma );
-
-	double TiOld = plasmaPtr->IonTemperature;
-	double TeOld = plasmaPtr->ElectronTemperature;
-
-	if ( ION_TEMPERATURE( u ) < 0.0 ) {
-#if defined( DEBUG )
-		std::cerr << "Error in SUNDIALS solve, due to negative ion temperature" << std::endl;
-#endif
-		return 1;
-	}
-	if ( ELECTRON_TEMPERATURE( u ) < 0.0 ) {
-#if defined( DEBUG )
-		std::cerr << "Error in SUNDIALS solve, due to negative electron temperature" << std::endl;
-#endif
-		return 2;
-	}
-
-	// We're now evolving the voltage with time 
-	plasmaPtr->IonTemperature = ION_TEMPERATURE( u );
-	plasmaPtr->ElectronTemperature = ELECTRON_TEMPERATURE( u );
-	plasmaPtr->ImposedVoltage = VOLTAGE( u );
-
-
-	
-	try {
-		plasmaPtr->SetTime(t);
-	} catch ( std::domain_error &e ) {
-		// Timestep too long?
-#ifdef DEBUG
-		std::cerr << "Evaluating RHS at t = " << std::setprecision( 20 ) << t << " ?!" << std::endl;
-#endif
-		return 3;
-	}
-	plasmaPtr->SetMachFromVoltage();
-	plasmaPtr->ComputeSteadyStateNeutrals();
-#if defined( DEBUG ) && defined( SUNDIALS_DEBUG ) && defined( INTERNAL_RK_DEBUG )
-	std::cerr << "t = " << t << " ; T_i = " << plasmaPtr->IonTemperature << " ; T_e = " << plasmaPtr->ElectronTemperature << " MachNumber " << plasmaPtr->MachNumber << std::endl;
-#endif
-
-	// I d omega / dt = ( Change in Angular Momentum )
-	// I d omega / dt = ( MomentumToVoltage )^(-1) dV/dt
-	// so dV/dt = MtoV * ( Change in Angular Momentum )
-	
-	double MomentumToVoltage = plasmaPtr->PlasmaColumnWidth * plasmaPtr->PlasmaCentralRadius() * plasmaPtr->CentralCellFieldStrength / plasmaPtr->MomentOfInertia();
-
-	try {
-		double IonHeating  = plasmaPtr->IonHeating();
-		double IonHeatLoss = plasmaPtr->IonHeatLosses();
-		double ElectronHeating  = plasmaPtr->ElectronHeating();
-		double ElectronHeatLoss = plasmaPtr->ElectronHeatLosses();
-
-#if defined( DEBUG ) && defined( SUNDIALS_DEBUG ) && defined( INTERNAL_RK_DEBUG )
-		std::cerr << " Ion Heating      = " << IonHeating      << " ; Ion Heat Loss       = " << IonHeatLoss      << std::endl;
-		std::cerr << " Electron Heating = " << ElectronHeating << " ; Electron Heat Loss  = " << ElectronHeatLoss << std::endl;
-#endif
-
-		ION_HEAT_BALANCE( uDot )      = ( IonHeating - IonHeatLoss );
-		ELECTRON_HEAT_BALANCE( uDot ) = ( ElectronHeating - ElectronHeatLoss );
-
-		// Should be negative to decelerate the plasma
-		double RadialCurrent = -VOLTAGE( u ) / plasmaPtr->ExternalResistance;
-		double AngularMomentumInjection = plasmaPtr->InjectedTorque( RadialCurrent );
-		double AngularMomentumLoss = plasmaPtr->TotalAngularMomentumLosses();
-
-		MOMENTUM_BALANCE( uDot ) = ( MomentumToVoltage ) * ( AngularMomentumInjection - AngularMomentumLoss );
-
-
-	} catch ( std::exception& e ) {
-		return -1;
-	} 
-
-	plasmaPtr->IonTemperature = TiOld;
-	plasmaPtr->ElectronTemperature = TeOld;
-	plasmaPtr->SetMachFromVoltage();
-	plasmaPtr->ComputeSteadyStateNeutrals();
-
-	return ARK_SUCCESS;
-}
-
 // In this mode, Mach Number is u(0) and T_i is u(1)
 // but we still solve both the power-balance equations by running to steady state
 int ARKStep_FixedTeSolve( realtype t, N_Vector u, N_Vector F, void* voidPlasma )
@@ -241,7 +124,6 @@ void MCTransConfig::doTempSolve( MirrorPlasma& plasma ) const
 	std::cerr << "Setting T_0 = " << InitialTemperature  << " to have M at t=0 fixed to " << InitialMach << std::endl;
 #endif
 
-
 	ION_TEMPERATURE( initialCondition ) = InitialTemperature;
 	ELECTRON_TEMPERATURE( initialCondition ) = InitialTemperature;
 	
@@ -249,8 +131,6 @@ void MCTransConfig::doTempSolve( MirrorPlasma& plasma ) const
 	plasma.IonTemperature = InitialTemperature;
 	plasma.SetMachFromVoltage();
 	plasma.ComputeSteadyStateNeutrals();
-
-	plasma.InitialiseNetCDF();
 
 	realtype t0 = 0;
 
@@ -323,7 +203,7 @@ void MCTransConfig::doTempSolve( MirrorPlasma& plasma ) const
 		plasma.IonTemperature = ION_TEMPERATURE( initialCondition );
 		plasma.SetMachFromVoltage();
 		plasma.ComputeSteadyStateNeutrals();
-		plasma.WriteTimeslice( t );
+		plasma.WriteTimeslice( tRet );
 #if defined( DEBUG )
 		std::cerr << "Writing timeslice at t = " << t << std::endl;
 #endif
@@ -342,8 +222,9 @@ void MCTransConfig::doTempSolve( MirrorPlasma& plasma ) const
 		     RelativeElectronRate < plasma.RateThreshold )
 		{
 #if defined( DEBUG )
-	std::cerr << "Steady state reached at time " << t << " with T_i = " << ION_TEMPERATURE( initialCondition ) << " ; T_e = " << ELECTRON_TEMPERATURE( initialCondition ) << std::endl;
+	std::cerr << "Steady state reached at time " << tRet << " with T_i = " << ION_TEMPERATURE( initialCondition ) << " ; T_e = " << ELECTRON_TEMPERATURE( initialCondition ) << std::endl;
 #endif
+			plasma.SetTime( tRet );
 			break;
 		}
 
@@ -353,9 +234,7 @@ void MCTransConfig::doTempSolve( MirrorPlasma& plasma ) const
 
 	plasma.ElectronTemperature = ELECTRON_TEMPERATURE( initialCondition );
 	plasma.IonTemperature      =      ION_TEMPERATURE( initialCondition );
-	// plasma.SetTime( EndTime );
 	
-	plasma.FinaliseNetCDF();
 
 #ifdef DEBUG
 	long nSteps = 0,nfeEvals = 0,nfiEvals = 0;
@@ -443,111 +322,3 @@ void MCTransConfig::doFixedTeSolve( MirrorPlasma& plasma ) const
 	plasma.ComputeSteadyStateNeutrals();
 }
 
-// Let the plasma decay through a resistor
-void MCTransConfig::doFreeWheel( MirrorPlasma& plasma ) const
-{
-	sundials::Context sunctx;	
-	sunindextype NDims = N_DIMENSIONS;
-	N_Vector initialCondition = N_VNew_Serial( NDims, sunctx );
-
-	double InitialTemperature = plasma.InitialTemp;
-	ION_TEMPERATURE( initialCondition ) = InitialTemperature;
-	ELECTRON_TEMPERATURE( initialCondition ) = InitialTemperature;
-
-	plasma.ElectronTemperature = InitialTemperature;
-	plasma.IonTemperature = InitialTemperature;
-	plasma.MachNumber = plasma.InitialMach;
-
-	plasma.ImposedVoltage = plasma.MachNumber * plasma.SoundSpeed() * plasma.PlasmaColumnWidth * plasma.CentralCellFieldStrength;
-
-	plasma.ComputeSteadyStateNeutrals();
-
-	plasma.InitialiseNetCDF();
-
-	realtype t0 = 0;
-
-	void *arkMem = ARKStepCreate( nullptr, ARKStep_FreeWheel, t0, initialCondition, sunctx );
-
-	if ( arkMem == nullptr ) {
-		throw std::runtime_error( "Cannot allocate ARKStep Working Memory" );
-	}
-
-	// Dummy Jacobian, will be filled by ARKStep with finite-difference approximations
-	SUNMatrix       Jacobian = SUNDenseMatrix( NDims, NDims, sunctx );
-	// Small system, direct solve is fastest
-	SUNLinearSolver  LS = SUNLinSol_Dense( initialCondition, Jacobian, sunctx );
-
-	ArkodeErrorWrapper( ARKStepSetLinearSolver( arkMem, LS, Jacobian ), "ARKStepSetLinearSolver" );
-	
-	
-
-	double abstol = plasma.SundialsAbsTol;
-	double reltol = plasma.SundialsRelTol;
-
-#ifdef DEBUG
-	std::cerr << "Using SundialsAbsTol = " << abstol << " and SundialsRelTol = " << reltol << std::endl;
-#endif
-	ArkodeErrorWrapper( ARKStepSStolerances( arkMem, reltol, abstol ), "ARKStepSStolerances" );
-	ArkodeErrorWrapper( ARKStepSetTableNum( arkMem, IRK_SCHEME, ARKSTEP_NULL_STEPPER ), "ARKStepSetTableNum" );
-	
-	ArkodeErrorWrapper( ARKStepSetUserData( arkMem, reinterpret_cast<void*>( &plasma ) ), "ARKStepSetUserData" );
-
-
-	const unsigned long MaxSteps = 1e4;
-	ArkodeErrorWrapper( ARKStepSetMaxNumSteps( arkMem, MaxSteps ), "ARKStepSetMaxNumSteps" );
-
-	realtype t,tRet = 0;	
-	int errorFlag;
-
-#ifdef DEBUG
-	std::cerr << "Solving from t = 0 to t = " << EndTime << std::endl;
-	std::cerr << "Writing output every " << OutputDeltaT << std::endl;
-#endif 
-	ArkodeErrorWrapper( ARKStepSetStopTime( arkMem, EndTime ), "ARKStepSetStopTime" );
-	for ( t = OutputDeltaT; t < EndTime; t += OutputDeltaT )
-	{
-		errorFlag = ARKStepEvolve( arkMem, t, initialCondition, &tRet, ARK_NORMAL );
-		switch ( errorFlag ) {
-			case ARK_SUCCESS:
-				break;
-			default:
-				throw std::runtime_error( "ARKStep failed with error " + std::to_string( errorFlag ) );
-			break;
-		}
-
-		plasma.ElectronTemperature = ELECTRON_TEMPERATURE( initialCondition );
-		plasma.IonTemperature = ION_TEMPERATURE( initialCondition );
-		plasma.SetMachFromVoltage();
-		plasma.ComputeSteadyStateNeutrals();
-		plasma.WriteTimeslice( t );
-#if defined( DEBUG )
-		std::cerr << "Writing timeslice at t = " << t << std::endl;
-#endif
-#if defined( DEBUG ) && defined( SUNDIALS_DEBUG )
-	std::cerr << "After evolving to " << tRet << " T_i = " << ION_TEMPERATURE( initialCondition ) << " ; T_e = " << ELECTRON_TEMPERATURE( initialCondition ) << std::endl;
-#endif
-	}
-
-	// We've solved and found the answer. Update the plasma object
-
-	plasma.ElectronTemperature = ELECTRON_TEMPERATURE( initialCondition );
-	plasma.IonTemperature      =      ION_TEMPERATURE( initialCondition );
-	plasma.SetTime( t );
-
-#ifdef DEBUG
-	long nSteps = 0,nfeEvals = 0,nfiEvals = 0;
-	ArkodeErrorWrapper( ARKStepGetNumSteps( arkMem, &nSteps ), "ARKGetNumSteps" );
-	ArkodeErrorWrapper( ARKStepGetNumRhsEvals( arkMem, &nfeEvals, &nfiEvals ), "ARKGetNumRhsEvals" );
-	std::cerr << "SUNDIALS Timestepping took " << nSteps << " internal timesteps resulting in " << nfiEvals << " implicit function evaluations" << std::endl;
-#endif 
-	// Teardown 
-	{
-		SUNLinSolFree( LS );
-		SUNMatDestroy( Jacobian );
-		N_VDestroy( initialCondition );
-		ARKStepFree( &arkMem );
-	}
-
-	plasma.SetMachFromVoltage();
-	plasma.ComputeSteadyStateNeutrals();
-}
